Move lifecycle functions to separate package
Move files in pkg/library to pkg/library/lifecycle to make adding more
library functions cleaner.
