+++ pkg/library/lifecycle/lifecycle_test.go
@@ -10,7 +10,7 @@
 //   Red Hat, Inc. - initial API and implementation
 //
 
-package library
+package lifecycle
 
 import (
 	"fmt"
