+++ pkg/library/lifecycle/testdata/lifecycle/init_and_main_container.yaml
