+++ pkg/library/lifecycle/testdata/lifecycle/no_events.yaml
